Make block-triggers scope to the task they are attached to.
Prior to this, if you defined (e.g.) a before task using a block,
like this:

  before :foo do
    puts "current task is #{current_task.name}"
  end

And then, you invoked :foo like this:

  task :bar do
    run "something"
    foo
  end

You'd see the before-foo block print "bar" as the name of the
current task. This meant that any commands run in the foo before
block would be scoped to the servers of :bar, when it really
makes more sense to scope them to the servers of :foo. This
commit makes it so that the before-foo block, above, would show
"foo" as the name of the current task, rather than "bar".