Release v2.12.0
This release revserts the very verbose logging introduced in the
previous version, it also
enables a handful of power-user features which are largely
un-documented, but shouldn't be
important unless you are looking for them. Undocumented code shouldn't
scare you, simply
read through deploy.rb in the Gem if you want to know how a new feature
works!

 * Update mapped commands to remove symlink deprecation warning. Despo
 * Pentara (despo)
 * Add the "rpm" remote dependency. Nick Hoffman (nickhoffman)
 * Add commented deploy:cleanup task to default recipe. Jean-Philippe
 * Doyle (j15e)
 * Teach deploy:web:enable to fail gracefully. Lee Marlow (lmarlow)
 * Ticket 193 alias task show wrong name when it is not overridden. Rafa
 * García (rgo)
 * Allow configuration of which roles assets are precompiled on.
 * Frederick Cheung (fcheung)
 * Fix transfer action to honor dry-run flag. Serg Podtynnyi (shtirlic)
 * Changed single to double quotes for Windows, fixes a Windows bug in
 * the HG module. Matthew J Morrison (mattjmorrison)
 * Add UnsharedRemoteCache (copied from eycap gem). Ben Symonds
 * (bensymonds)

As ever, a sincere thanks to all contributors, and do not hesitate to
contact me if this
release causes problems for you.