Fix current_release for after rollback case
Fixes capistrano/capistrano#154 such that instead of getting the newest
release directory it will follow the `current_path` symlink and read the
path of the linked to path. Even if a failed deploy release directory is
not cleaned up afterwards, the `current_release` variable should point to
the correct path still and avoids using the failed deploy release on
subsequent Capistrano runs until the next successful deploy.

Should still work when `current_dir` symlink doesn't exist as well as it
falls back to `release_path` if `current_path` does not exist to read its
link path.