lib/capistrano/processable.rb: log output if nothing happens for a while
Currently, cap can hang indefinitely waiting for output from a server.
This adds a 10 second timeout to the IO.select command and prints the
names of the servers to which we still have open channels.  This allows
us to easily see which servers are taking a long time or hanging.