Raise error when more than one asset manifest file is detected
If a user is upgrading from Rails 3 to 4, the assets:precompile task will ignore the manifest.yml
that was generated in Rails 3 and generate its own JSON manifest. It's impossible for us to know
which manifest the user's current code requires, so it's important that we only have one manifest
in the assets directory at a time. When we detect two manifests, we raise an error pointing the
user towards a wiki entry explaining the problem and instructing them to remove the Rails 3.x manifest.