Improve selection of primary host for a role.
Prior to this commit the "primary" server in each role was considered to be
the first one defined. This commit extends the behaviour of primary selection
to look to see if any of the hosts have the `primary` property set to a truthy
value.

In case that more than one server has a truthy `primary` property the first
one is taken as previously.

In order to set the primary property the extended `server()` syntax (that is
not `role()`) must be used, see b56206e.