`--version` Prints Both Cap and Rake versions.
Small modification, shows a proof of concept on how we can enable other
top-level command line flags by interacting with Rake (they could make this
easier, to be honest).

Two pending tests for a `--trace` option and a `--format` option, I'm not
fixed on these names, but the notion stands.