Workaround a short-sighted issue with roles.
Deploying an environment that has no :web, or :app roles, the role is
:workers, but the environment should version, and symlink the worker code,
just as it would a web, or app server.

The same might be true of any role.

The prevailing error message is:

    cap aborted!
    Your filter `(no filter)` would remove all matching servers
    .../gems/capistrano-3.0.0.pre9/lib/capistrano/configuration/servers.rb:73:in `filtered_servers'
    .../gems/capistrano-3.0.0.pre9/lib/capistrano/configuration/servers.rb:53:in `filter'
    .../gems/capistrano-3.0.0.pre9/lib/capistrano/configuration/servers.rb:48:in `fetch_roles'
    .../gems/capistrano-3.0.0.pre9/lib/capistrano/configuration/servers.rb:17:in `roles_for'
    .../gems/capistrano-3.0.0.pre9/lib/capistrano/configuration.rb:41:in `roles_for`

This patch works around the issue by defining that symlink works on :all
roles, which is inkeeping with what the rest of the code is doing.

This resurects the logic behind the `:no_release => true` option that could be
set on servers in previous versions of Cap. I think in this case, it wouldn't
be unreasonable to filter most of `deploy.rake` with `reject: :no_deploy` or
similar, and set that as an option in the config file template that is
generated, for the :db role, and write a line or two about how that works.