Symlink tasks should happen on all roles, not just app
When deploying to multiple different roles, symlinks currently only
happen on the app role. When you then try and precompile assets on the
web role for example, the symlinks that it requires to link to certain
config files may not be there. This fix changes the default symlink
behaviour to happen on _all_ roles as this seems like a more sensible
default.