use latest_release instead of current_release in assets.rb
On the first deploy, current_release doesn't seem to be set, so it tries to copy the assets_manifest to /assets_manifest.json, which it doesn't have permission to do. Switching it to latest_release seems to fix that, but I'm not sure if that's correct in all other cases.