don't check git connection
It may not be required or allowed to have a connection to the git
server from the app host. When using capistrano-rsync it isn't even required to have git
installed. I don't see why this is required. Perhaps the check can be
moved to just before the create_release command, but I don't see the
point in doing that.