Update the configuration for Vagrant 1.7+
Vagrant 1.7+ defaults to removing the insecure SSH key which causes
the cap tasks to fail. Add configuration to use old method

Requires that the Vagrant configuration syntax is updated, so also
update the box/url and network definitions

This means the latest version of Vagrant will be needed to run the
test features.