Introduce banner for Harrow.io
Sorry chaps, we're trying a push to promote Harrow.io to try and make my
work on Capistrano sustainable, and to promote our new SaaS Harrow.io,
which is a new take on CI tooling focused on collaboration.

It was designed with "Web Based Capistrano" in mind, but as you might
imagine, once one builds a tool for running Cap in the "cloud" it's not
a big leap to run anything else.

Ping me off-list if you have any issues with anything here, we tried to
be very fair with the behaviour, cookie lifetimes and positioning.

tl;dr:

- Invisible by default (display none)
- Reasonable cookie lifetime
- Made visible using jQuery (.css("display", "block")