Make path to git wrapper script configurable
The wrapper script is known to cause problems on multi-user deployment
systems, because the script is uploaded to the same location within
tmp_dir, resulting in permissions issues when multiple users attempt to
write to the same file, which is (sensibly) only writeable by its owner.

Using the application name in the path goes some way to addressing this,
but doesn't help when multiple users deploy the same application.

Making the path to the wrapper script configurable makes it easier for
developers and sysadmins to customise the configuration in a way that
makes sense for their systems, without needing to override the default
task, create tmp dirs for individual users or some other workaround that
requires background knowledge of the internal workings of Capistrano. It
also allows the path to be referenced in any custom cleanup tasks :-)

Conflicts:
	CHANGELOG.md
	lib/capistrano/tasks/git.rake