Deduplicate list of linked directories
With this, we ensure to only handle each linked directory once.
Also, if a linked file is in a directory that is already being
handled, there is no need to link the same directory twice.

This should help to keep commands shorter, which might be an issue on
some shells. It certainly is easier on the eyes and logs.