Normalize before/after task defn. and invocation
Capistrano and Rake both have ways of enhancing existing tasks with additional
actions or prerequisites, but the behavior is subtly different.

In Capistrano, before/after implicitly create a new named task (when a block is
supplied). When this feature is used, task creation and task invocation take
place in two different runtime contexts, which leads to subtle inconsistencies
with regards to namespaces.

Rake's solution is to *always* search the task's namespace when looking up a
prerequisite. E.g. if we add `foo` as a prerequisite like this:

`Rake::Task['deploy:starting']`.enhance([:foo])

Then during task invocation, Rake will look for `foo` in the `deploy` namespace.
If it can't find `deploy:foo`, then it falls back to looking for `foo`. (I find
this behavior non-intuitive, but that's how it works.)

This is at odds with how Capistrano's system behaves. If we try to accomplish
the same thing using Capistrano `before` block syntax:

before('deploy:starting', :foo) do
  # ...
end

This creates a root-level `foo` task, but then since Capistrano uses Rake's
prerequisite system under the hood, at invocation time `deploy:foo` is invoked.
The task being declared and the lookup being performed don't match.

This commit fixes this inconsistency. Now `foo` is created at the root level as
it always has been, but the invocation behavior is changed: Capistrano now
correctly invokes the root-level `foo` instead of searching the parent tasks's
namespace.

Accomplishing this required some tricky code involving Rake's special `rake:`
namespace to force the task lookup behavior we want, and definitely goes against
the grain of Rake. However this seemed to be the best solution that makes the
task declaration and invocation behavior consistent when `before` and `after`
are used.

No existing tests broke as a result of this change, which speaks well for
backwards compatibility. I added more tests to cover the edge cases that are
impacted.