Move plugin installation logic out of initializer
Plugins are now installed via a `install_plugin` DSL method, rather than as a
side-effect of calling `new`. You may pass either the plugin class or the
instance of the class to `install_plugin`; both accomplish the same thing.

The logic for executing the plugin is now encapsulated in the following class:
`Capistrano::Configuration::PluginInstaller`.