Add deferred validation of proc values
This fixes a bug where the `:application` validation rule would fail if
given a proc value.

The fix involves changing how validation rules are applied to procs.
For a proc, instead of validating immediately, we defer validation until
when the proc is fetched and called.

Since this causes the validation logic to be non-trivial, I factored it
out into a separate class using the decorator pattern. Now `Variables`
handles the core set/fetch logic, and `ValidatedVariables` decorates it
to layer on validation.