Make all SCMs inherit from SCM::Plugin
This makes it much easier (no hacks needed) to determine whether the
user has installed an SCM plugin: simply check that the plugin responds
to `scm?` and that it returns true. A new SCM::Plugin base class makes
it easy to write such a plugin.

This commit also updates the website docs that explain how to write SCM
plugins so that they refer to this new technique.