Include link to DNSimple in the footer. They sponsor our DNS.
Whilst trying to reduce the impact the loss of any one maintainer can
have, we were offered a solution to the DNS ownership problem by
DNSimple who kindly sponsored our DNS entry. This allows the three core
maintainers to have access to the DNS system incase changes need to be
made.

Part of the deal was including a link to DNSimple in the footer of the
site (as Hanamai, and others do).

I forgot, apparently to keep my end of the bargain, and during a recent
audit they've checked in with me to see if we're still making use of the
sponsored service. As I like DNSimple and find them to be very useful,
I'd like them to continue to sponsor us, and offer this tweak to the
doce to fulfil our obligations in that regard.