Invoke `bundle` correctly within feature tests
During the feature tests we create a separate directory with its own
`Gemfile` and use `bundle` and `bundle exec` within that directory.

However if `bundle exec` was used to run the feature tests, then a
bundler environment is already loaded, and so the test Gemfile is never
consulted. Running `bundle` doesn't have the desired effect, because
the child process inherits the bundler environment variables from the
parent process.

Bundler's solution for this problem is the `Bundle.with_clean_env`
method. This ensures that the child process doesn't inherit the parent's
bundler environment.