Allow `cap -T` to work when no Capfile is present
Previously, the `cap -T` command assumed the presence of the
`load:defaults` task. This task is only present if a proper Capfile is
loaded, and so `cap -T` wouldn't work in a new, not yet "capified"
project.

This commit fixes the bug by testing that `load:defaults` is defined
before calling it.