Auto-bump version number in generated release notes
We are using Release Drafter[^1] to automatically write our release
notes.

Before, Release Drafter was configured to increment the version number
at the patch level, regardless of what was included in the release
notes. For example, if the current version is 3.17.3, the draft of the
next release would assume a version number of 3.17.4.

With this commit, Release Drafter will now bump the version number based
on the concents of the release. A patch bump is the default, but if
there is a new feature or breaking change in the release, it will now
recommend a minor bump.

This commit also modifies the release notes template to add a link to
the full diff of the release on GitHub.

[^1]: https://github.com/release-drafter/release-drafter