Upgrade to rubocop 1.81.7; generate `.rubocop_todo.yml` (#2172)
Before, we were using a very old version of RuboCop that did not work
well on modern Ruby. This made for a poor contributor experience,
because the only way to be sure code met our style guidelines was to
push to CI and wait for the results.

This PR upgrades RuboCop to the latest version (1.81.7) and makes it a
standard part of the Gemfile and default rake task. Now contributors can
just run `bundle exec rake` and get immediate feedback on unit tests and
style checks.

The latest RuboCop brings some new rules (and renamed rules) compared to
the old version we were using previously. I disabled some of the new
rules that are not relevant. For the remaining ones, I generated a
`.rubocop_todo.yml` with the violations. This keeps this PR free of any
production code changes that might introduce bugs. We can decide later
if/how to address the todo items.